+++ src/__init__.py
@@ -3,6 +3,7 @@
 from flask import Flask
 from werkzeug.middleware.proxy_fix import ProxyFix
 
+import src.topics.routes
 import src.users.routes
 from src import routes
 from src.config import Config
@@ -10,6 +11,7 @@ from src.config import Config
 app = Flask(__name__)
 app.register_blueprint(routes.bp)
 app.register_blueprint(src.users.routes.bp)
+app.register_blueprint(src.topics.routes.bp)
 app.config.from_object(Config)
