+++ src/topics/__init__.py
@@ -1,0 +1,1 @@
+"""'topics' package's initialization module."""
