+++ src/topics/templates/topics.html
@@ -1,11 +1,11 @@
 {% extends "base.html" %}
 
 {% block content %}
-    <h1><a href="{{ url_for('routes.create_topic') }}">Create topic</a></h1>
+    <h1><a href="{{ url_for('topics.create_topic') }}">Create topic</a></h1>
     {% for topic in topic_list %}
     <br>
         <ul>
-            <li><b><a href="{{ url_for('routes.topic_page', topic_id=topic.id) }}">{{ topic.title }}</a></b></li>
+            <li><b><a href="{{ url_for('topics.topic_page', topic_id=topic.id) }}">{{ topic.title }}</a></b></li>
             <li>{{ topic.description }}</li>
             <li>{{ topic.author.username }}</li>
             <li>at {{ topic.created_at }}</li>
